+++ src/com/techfiesta/asaan/utility/Constants.java
@@ -14,7 +14,7 @@ public class Constants {
 	public static final Long ROW_TYPE_SUBMENU = 1L;
 	public static final Long ROW_TYPE_MENUITEM = 2L;
 
-	public static final long MENU_TYPE_DINE_IN = 0L; // 0 = dine-in, 1 =
+	public static final int MENU_TYPE_DINE_IN = 0; // 0 = dine-in, 1 =
 														// carry-out
 	public static final long MENU_TYPE_CARRYOUT_DELIVERY = 1L; // 0 = dine-in, 1
 																// = carry-out
